--- src/scitex/repro_rng
@@ -1,1 +1,0 @@
-rng
